+++ src/main/java/com/ead/gearup/dto/user/PasswordChangeRequest.java
@@ -1,16 +1,20 @@
-package com.ead.gearup.dto;
+package com.ead.gearup.dto.user;
 
+import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
 import lombok.Data;
 
 @Data
-public class ChangePasswordRequest {
+public class PasswordChangeRequest {
     @NotBlank(message = "Current password is required")
     private String currentPassword;
 
-    @NotBlank(message = "New password is required")
+    @NotBlank(message = "Password is required")
     @Size(min = 8, message = "Password must be at least 8 characters long")
+    @Pattern(regexp = "^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[^a-zA-Z0-9]).*$", message = "Password must contain uppercase, lowercase, number, and special character")
+    @Schema(example = "string")
     private String newPassword;
 
     @NotBlank(message = "Password confirmation is required")
